+++ src/Raven.CodeAnalysis/BoundTree/BoundBreakStatement.cs
@@ -2,4 +2,10 @@ namespace Raven.CodeAnalysis;
 
 internal partial class BoundBreakStatement : BoundStatement
 {
+    public BoundBreakStatement(ILabelSymbol? targetLabel = null)
+    {
+        TargetLabel = targetLabel;
+    }
+
+    public ILabelSymbol? TargetLabel { get; }
 }