+++ src/Raven.CodeAnalysis/BoundTree/BoundContinueStatement.cs
@@ -2,4 +2,10 @@ namespace Raven.CodeAnalysis;
 
 internal partial class BoundContinueStatement : BoundStatement
 {
+    public BoundContinueStatement(ILabelSymbol? targetLabel = null)
+    {
+        TargetLabel = targetLabel;
+    }
+
+    public ILabelSymbol? TargetLabel { get; }
 }